pyproject.toml
[tool.poetry]
name = "erebus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
imgurpython = "^1.1.7"
aiohttp = "^3.7.4"
discord-buttons = "^0.5"
discord-components = "^0.5.2"
Pillow = "^8.2.0"
Pillow-PIL = "^0.1dev"
python-resize-image = "^1.1.19"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
